backend-server/requirements.txt
amqp==5.1.1
aniso8601==9.0.1
bcrypt==4.0.1
billiard==4.1.0
blinker==1.6.2
cachelib==0.9.0
celery==5.3.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.0
dnspython==2.4.2
email-validator==2.0.0.post2
eventlet==0.33.3
Flask==2.3.3
Flask-Bcrypt==1.0.1
flask-blueprint==1.3.0
Flask-Caching==2.1.0
Flask-Cors==4.0.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
flask-marshmallow==0.15.0
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.3.0
Flask-SQLAlchemy==3.1.0
Flask-WTF==1.1.1
fonttools==4.43.0
greenlet==2.0.2
idna==3.4
importlib-resources==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
kombu==5.3.2
MarkupSafe==2.1.3
matplotlib==3.8.0
numpy==1.26.0
packaging==23.2
pandas==2.1.1
passlib==1.7.4
pdfkit==1.0.0
Pillow==10.0.0
prompt-toolkit==3.0.39
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
redis==4.6.0
six==1.16.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
tzdata==2023.3
vine==5.0.0
waitress==2.1.2
wcwidth==0.2.6
Werkzeug==2.3.7
WTForms==3.0.1
